Add support for reading GuideStar objIds as 64-bit
Also added code to allow unit test to run standalone for testing.

+++ python/pfs/datamodel/guideStars.py
@@ -174,7 +174,7 @@ class GuideStars:
         """
         hdu = fits[cls._hduName]
 
-        return cls(hdu.data["objId"].astype(np.int32),
+        return cls(hdu.data["objId"].astype(np.int64),
                    hdu.data["epoch"],
                    hdu.data["ra"].astype(np.float64),
                    hdu.data["dec"].astype(np.float64),
